+++ Tests/AdvancedMappingCrud.Repositories.Tests/AdvancedMappingCrud.Repositories.Tests.Application/Interfaces/ServiceToServiceInvocation/IStoredProcService.cs
@@ -1,0 +1,16 @@
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using AdvancedMappingCrud.Repositories.Tests.Application.ServiceToServiceInvocation;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.Contracts.ServiceContract", Version = "1.0")]
+
+namespace AdvancedMappingCrud.Repositories.Tests.Application.Interfaces.ServiceToServiceInvocation
+{
+    public interface IStoredProcService
+    {
+        Task<List<GetDataEntryDto>> GetData(CancellationToken cancellationToken = default);
+    }
+}